Fixed compatibility issues with Python 3

+++ python/pyramid.py
@@ -38,7 +38,7 @@ class BurtOF:
 
     def conv2SepMatlab(self, I, fen):
 
-        rad = (fen.size-1)/2
+        rad = int((fen.size-1)/2)
         ligne = np.zeros((rad, I.shape[1]))
         I = np.append(ligne, I, axis=0)
         I = np.append(I, ligne, axis=0)
